fix: move some layer2 utils to correct folder
