--- packages/shared/lib/core/layer-2/actions/index.ts
@@ -1,2 +1,0 @@
-export * from './getLayer2MetadataForTransfer'
-export * from './parseLayer2MetadataForTransfer'
