+++ packages/shared/lib/core/layer-2/interfaces/index.ts
@@ -1,3 +1,4 @@
+export * from './layer2-allowance.interface'
 export * from './layer2-parameters.interface'
 export * from './layer2-smart-contract-call-data.interface'
 export * from './layer2-transfer-allowance-metadata.interface'
