+++ packages/shared/lib/core/layer-2/interfaces/layer2-smart-contract-call-data.interface.ts
@@ -1,4 +1,4 @@
-import { ILayer2Allowance } from '@core/layer-2/interfaces/layer2-allowance.interface'
+import { ILayer2Allowance } from '@core/layer-2/interfaces'
 
 export interface ILayer2SmartContractCallData extends ILayer2Allowance {
     senderContract: string
