+++ packages/shared/lib/core/layer-2/utils/index.ts
@@ -1,4 +1,6 @@
 export * from './addGasBudget'
 export * from './getDestinationNetworkFromAddress'
+export * from './getLayer2MetadataForTransfer'
 export * from './isLayer1Destination'
 export * from './parseLayer2Metadata'
+export * from './parseLayer2MetadataForTransfer'
