+++ packages/shared/lib/core/wallet/utils/generateActivity/helper/getMetadataFromOutput.ts
@@ -1,7 +1,7 @@
 import type { IMetadataFeature } from '@iota/types'
 import { isParticipationOutput } from '@contexts/governance/utils'
 import { EXTERNALLY_OWNED_ACCOUNT } from '@core/layer-2/constants'
-import { parseLayer2MetadataForTransfer } from '@core/layer-2/actions'
+import { parseLayer2MetadataForTransfer } from '@core/layer-2/utils'
 import { containsControlCharacters, Converter } from '@core/utils'
 import type { Output } from '@core/wallet/types'
